rebases tests from js to tsx with fix
