+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "description": "Личный проект «Что посмотреть» от HTML Academy",
   "scripts": {
-    "eslint": "eslint src/**/*.js src/**/*.jsx",
+    "eslint": "eslint src/**/*.js",
     "test": "npm run eslint && jest",
     "build": "webpack --mode development",
     "start": "webpack-dev-server"
@@ -13,6 +13,7 @@
     "@babel/core": "7.4.4",
     "@babel/preset-env": "7.4.4",
     "@babel/preset-react": "7.0.0",
+    "@types/jest": "24.0.14",
     "@types/react": "16.8.19",
     "axios-mock-adapter": "1.16.0",
     "babel-jest": "24.8.0",
@@ -24,6 +25,7 @@
     "eslint-plugin-react": "7.13.0",
     "jest": "24.8.0",
     "react-test-renderer": "16.8.6",
+    "ts-jest": "24.0.2",
     "ts-loader": "6.0.2",
     "typescript": "3.5.1",
     "utility-types": "3.7.0",
