+++ src/components/message/message.test.tsx
@@ -1,0 +1,14 @@
+import * as React from 'react';
+import * as renderer from 'react-test-renderer';
+import Message from './message';
+
+it(`Message correctly render text:`, () => {
+  const tree = renderer.create(
+    <Message
+      text={`is correct message`}
+      style={``}
+    />
+  ).toJSON();
+
+  expect(tree).toMatchSnapshot();
+});
