+++ src/components/sign-in/sign-in.test.tsx
@@ -1,14 +1,14 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-import SignIn from '../sign-in/sign-in.jsx';
+import * as React from 'react';
+import * as renderer from 'react-test-renderer';
+import SignIn from './sign-in';
 
 it(`SignIn correctly renders with message: \'Sign in here\' text props: `, () => {
   const mockHandle = jest.fn();
   const tree = renderer.create(
-      <SignIn
-        onLogin={mockHandle}
-        message={`Sign in here`}
-      />
+    <SignIn
+      onLogin={mockHandle}
+      message={`Sign in here`}
+    />
   ).toJSON();
 
   expect(tree).toMatchSnapshot();
@@ -17,10 +17,10 @@ it(`SignIn correctly renders with message: \'Sign in here\' text props: `, () =>
 it(`SignIn correctly renders without message: `, () => {
   const mockHandle = jest.fn();
   const tree = renderer.create(
-      <SignIn
-        onLogin={mockHandle}
-        message={``}
-      />
+    <SignIn
+      onLogin={mockHandle}
+      message={``}
+    />
   ).toJSON();
 
   expect(tree).toMatchSnapshot();
