+++ src/components/small-movie-card/small-movie-card.e2e.test.tsx
@@ -1,10 +1,10 @@
-import React from 'react';
-import Adapter from 'enzyme-adapter-react-16';
-import Enzyme, {shallow} from 'enzyme';
-import SmallMovieCard from '../small-movie-card/small-movie-card.jsx';
+import * as React from 'react';
+import * as Adapter from 'enzyme-adapter-react-16';
+import { configure, shallow } from 'enzyme';
+import SmallMovieCard from './small-movie-card';
 import TEST_MOCKS from '../../mocks/test-mocks.js';
 
-Enzyme.configure({adapter: new Adapter()});
+configure({ adapter: new Adapter() });
 
 const MOCK_ACTIVE = false;
